Implement `bundle install`
Runs `bundle install` and `bundle clean`. The behavior closely matches the existing v2 buildpack. One tweak is that we're detecting in the output of the `bundle install` includes `Fetching`. If not, then we can skip `bundle clean` (saves a bit of invocation and runtime). 

All behavior here is tested in a unit test that is fairly fast:

```
$ be rspec spec/unit/bundle_install_spec.rb
..

Finished in 10.23 seconds (files took 0.50482 seconds to load)
2 examples, 0 failures
```

The downside of the test is that it uses system bundler and system ruby, but that's fine as we're testing the `bundle install` logic and `bundle clean` logic in isolation from detecting=>downloading bundler.

The only thing missing from this implementation is cache support

## Tweaks

- Rename "user_output" to "user_comms"
- Add UserComms#warn_now and warn_later
- UserComms#error_and_exit does what it says it does
- Add EnvProxy.default for manipulating environment variables that the user can set/override
- EnvProxy::Base subclasses now return self when manipulating their values so we can chain commands
- EnvProxy::Base#to_env returns a string that can be used in the shell
- EnvProxy::Base#value returns the existing value of the env var that the proxy wraps
- CNB buildpack spec can now execute run commands in multiple threads
- CNB buildpack spec now correctly shell escapes command inputs for docker run commands
- Bash now supports Bash#stream output is streamed to a block so that the Bash class doesn't have to know about UserComms class.