Don't re-allocate Pathname
Calling Pathname.new creates a new pathname each time while calling `Pathname()` will only allocate it it's not already a pathname.

```
require 'benchmark/ips'

require 'pathname'

p = Pathname.new(".")
Benchmark.ips do |x|
  x.report("Pathname.new") { Pathname.new(p) }
  x.report("Pathname()  ") { Pathname(p) }
  x.compare!
end


# Warming up --------------------------------------
#         Pathname.new   198.032k i/100ms
#         Pathname()       1.546M i/100ms
# Calculating -------------------------------------
#         Pathname.new      1.952M (± 2.4%) i/s -      9.902M in   5.076761s
#         Pathname()       15.322M (± 2.5%) i/s -     77.314M in   5.049261s

# Comparison:
#         Pathname()  : 15321873.5 i/s
#         Pathname.new:  1951557.0 i/s - 7.85x  (± 0.00) slower
```

The resulting code is nearly ~7-8x faster. While it's likely not going to be a bottleneck anytime soon, let's not be in a habit of writing slow code when it's just as easy and clean to use faster code.