[close #9][changelog skip] Vendor TOML parsing lib
The previous TOML parsing library was small, but not robust. To properly parse TOML a parser is needed. The options are currently available at https://www.ruby-toolbox.com/search?q=toml.

The `tomlrb` was chosen to be vendored as it contains it's own parser while other popular TOML libraries depend on a parser (such as parselet  or citrus). 

I decided to manually install and require this library instead of depending on bundler so that we have more control over the process.

I also added a license to the "dumper" and moved it to it's own file.