Fix lockfile parser with test
The lock file parser has reference to other bundler defined constants such as Bundle::Source. While we could attempt to whack-a-mole each of these requires, that would be brittle. Instead we can require the entirety of bundler.

The existing test did not catch this because it is being called from within a bundler context already. To write a failing test that would have caught this issue, I needed a pristine ruby environment without bundler loaded. To get this I wrote the code to text against to a script file. Then I am manually executing that script via Bash.new("ruby script.rb"). And I'm wrapping that in a Bundler.with_original_env. This combination provides a similar/same environment as the Ruby buildpack (executing the code without bundler loaded yet).

This approach caused the old code to fail, and allows the new code to pass. It will hopefully be more robust then the previous lock file parser tests.