Detect execsjs and install node on V2
The `execjs` gem allows Ruby code to execute javascript. It does this by searching for several possible implementations:

- The ruby racer (V8 distributed as a Ruby gem)
  - This gems is extremely heavy, takes a long time to compile, frequently fails compilation and is generally not commonly used.
- MinRacer (A paired down interface to V8 distrbuted as a Ruby gem)
  - From the readme: Unlike therubyracer, mini_racer only implements a minimal bridge. This reduces the surface area making upgrading v8 much simpler and exhaustive testing simpler.
- System installation of `node`

Prior to the introduction of `mini_racer` the most common and popular development strategy was to depend on a system version of `node`. This PR implements existing behavior in the old buildpack where we install node if `execjs` is present in the gemfile.lock https://github.com/heroku/heroku-buildpack-ruby/blob/38cb9c1e8e16fd09cdf2086387fe9a1882d920de/lib/language_pack/ruby.rb#L1153-L1154.

Unlike the old Ruby buildpack we do not manually manage and install a node binary, instead we rely on the Node buildpack to install and manage node so that user expectations are more aligned.

## Implementation details

The Node.js buildpack requires a `package.json` copy at the root of the buildpack for detection. In our case the developer may not have added one, but we still need to use the `heroku/nodejs` buildpack to manage our node for us. To accomplish this the Ruby buildpack detects this scenario and writes out a minimal `package.json` so that the nodejs buildpack will not fail detection.

## CNB Concerns

This same strategy does not work for CNB because the detect phase runs in parallel so even if the Ruby buildpack writes a `package.json` to disk as part of it's detect phase, it's too late and the nodejs buildpack's detect has already fired. While this strategy seems to work well enough for the V2 interface, we need to figure out how to support this behavior with the CNB interface.