Support User Env config vars
Most other buildpacks import the user defined config values into the same process that is running the buildpack code. The Ruby buildpack gains additional control by not doing this and instead parsing and storing this information so it can be used later.

## UserEnvFromDir

This is a multi purpose class that encapsulates env parsing logic. It acts similar to the ENV hash.

Populating this object and creating it are decoupled. This allows us to create a global instance before loading the env directory, and then later populate it.

## UserEnv constant

The `UserEnv` constant is a global instance of `UserEnvFromDir`. All code can depend on it being present. It is being accessed by `Bash`, `EnvProxy::Default`, `EnvProxy::Base`, and `EnvProxy` as well as the top level HerokuBuildpackRuby methods

## EnvProxy::Default

This class is now aware of the UserEnv object. If a key is present in the UserEnv object it will prefer that value.

The big caveat here is that if someone is specifying a default env proxy object, they should be aware that it will mutate the buildpack's ENV. It's not ideal, but this prevents us from having multiple sources of truth in regards to the current value. Since this behavior is limited only to constants declared as EnvProxy.default, it limits the impact on the current process to a subset of values that the user is expected to be able to set anyway.

We could decouple this behavior in the future, but I think in the long run mutating ENV in this one sub case will cause fewer headaches.

## EnvProxy::Base

To support EnvProxy::Default behavior the base object now takes an optional `user_env:` kwarg and defaults to `UserEnv`.

## EnvProxy.default

To support building a EnvProxy::Default object, this method now accepts an optional `user_env` kwarg and defaults to `UserEnv`

## Bash.rb

The bash object is aware of the global `UserEnv` object and by default will run commands with the user's config values. By default bash commands will use the user's config vars, but this can be toggled off by passing in a falsey value:

```
Bash.rb("ls", user_env: false).run
```

## EnvProxy spec

To simplify tests, I've wrapped the specs in a top level module definition so that we don't have to use the fully qualified HerokuBuildapackRuby constant all over the place.

## Detect Ruby Version

The only place that we need to directly call the bootstrapped ruby version is to retrieve the `bundle platform --ruby` output. Since our version of Ruby is likely different that other versions of Ruby we will execute this without env vars. This allows us to not have to deny-list env vars such as RUBYOPT. I believe this is sufficient for our cases,