RubyVersion produces RubyVersionInfo instance
Previously we were returning a string of the Ruby version. That was fine when we only supported MRI, but now that I'm adding in JRuby support we need a more robust object.

This object wraps the return value so it can be queried for more information later. It also acts like a string so that anywhere we were using code like:

```
"ruby-#{version}.tgz"
```

It should work exactly the same.